refactor: migrate schemas dir to root
